accept strings or symbols as keys in opts hash
was confused for a while why specifying platforms manually wasn't
working before realizing that keys in the option hash needed to be
strings, and i was using the 2.0 hash literals. made a change here to
symbolize keys from metadata.json and in the opts being passed in so
either way can be used, enabling use of the new 2.0 hash literal format.

added a test using symbols instead of strings as keys in a hash passed
to on_supported_os.