Support specifying facter version
Before this commit rspec-puppet-facts would always construct a filter
based on the facter version that is loaded.

This commit adds support to #on_supported_os for passing :facterversion
with a specific facter version through the options hash.

If a :facterversion is passed, it will construct a statement expression
for the filter based off that.

If a :facterversion is not passed, it will default to the version of
facter that is loaded, staying compatible with the old behavior.