Search in both directions when looking for facts
The existing code only tried to downgrade the facter version when
looking for facts in FacterDB.  It could also get stuck doing this if
`facterversion` was given without a `revision` part.

With this commit, if the given facterversion contains no facts for the
specific OS, it will try to find the closest facter version present in
FacterDB.  It will try up to 10 versions in each direction (starting by
downgrading the version).

Warning messages are also tweaked to make it more obvious which facts
being requested were missing and which OSes are not going to be tested
etc.

Fixes #100