(PDK-1438) Auto select Facter version based on Puppet version
An endpoint has been added to the Forge API allowing us to pull down
a list of puppet-agent AIO package versions and the components inside
them. With this we can now automatically look up the used Puppet version
and select the appropriate Facter version to use when querying FacterDB.

When looking up the Puppet version, it will look backwards through the
known Puppet versions to find the next best match if there is not
a direct match. This means that when a new Puppet version is released,
it will select the Facter version for the latest version of Puppet that
it knows about, resulting in a pretty close match. If it searches
backwards and can not find any Puppet versions then it will default to
the currently loaded Facter version.

In order to ensure that rspec-puppet-facts continues to work offline,
the puppet-agent component data is fetched using the
`puppet_versions:update` task and committed to the repo.