Do not query for the exact facter version
This assumes that no exact match can be found in most cases. This saves
a call to Facterdb. Some testing on puppet-nginx this reduced the time
for a dry run by about 1 to 2 seconds, from about ~11 to ~9 seconds.

Actual results will vary on the FacterDB. Note that a new Facter version
will trigger the bad code unless all operating systems are updated so
this is a very likely code path.

It also makes the easier to read.