Do not query for the exact facter version
This assumes that no exact match can be found in most cases. This saves
a call to FacterDB. Some testing on puppet-nginx this reduced the time
for a dry run by about 1 to 2 seconds, from about ~11 to ~9 seconds.

Implementation wise this switches to using Gem::Requirement to detect
matching versions. This means you can also specify `~> 3.8` as a valid
facterversion, or `>= 3 < 5`.