(bug) - Prevent duplicated fact keys
Previously, when registering a custom fact with rspec-puppet-facts, the
fact key would be converted to a string. This lead to duplicate facts
being present in the facts hash, as default fact keys are stored as sym.

To fix this, we remove a `to_s` conversion on the key in the
register_custom_fact method, unless the rspec config stringify keys is true.