Use a single FacterDB.get_facts call
The major cost is FacterDB.get_facts. Instead of calling it once for
every OS version this instead calls it once with a complex filter. It
then uses JGrep to simulate another smaller DB and futher query that.

In my non-scientific test this reduces the load time of puppet-example
from 3.5 to 2.5 seconds.