(feat) - Add merge_facts option
This commit adds an option to the add_custom_fact method, which allows
the user to specify whether they wish to deep_merge facts.

This allows the user to determine whether they wish to completely
overwrite a fact, or simply merge values into the existing fact in a
given fact hash.