stringify fact keys by default
for a long long time, the first level of keys in a factsets were
symbols. Following nested hashes always had strings. This doesn't make
sense and looks confusing. Also the original data contains only strings,
we we actually iterated on the data and converted it to symbols. We now
flip the default. This saves performance and makes code more readable.