CI: Dont validate puppet_agent_facter_versions.json anymore
For a long time, we talked to the forge API to get a list of
puppet-agent releases and their components. One of the components is
facter. We generated a list with an agent<>facter mapping. This file
lives in ext/puppet_agent_facter_versions.json.

Now that we cannot test with puppet agent from Perforce, scraping that
list doesn't make sense.

From OpenVoxProject, we don't have a related API yet. So instead of
switching to a different API, we disable the CI check for now.